refactor: refactor filter of CloudService
