+++ src/spaceone/inventory/manager/cloud_service_manager.py
@@ -3,9 +3,12 @@ import copy
 from datetime import datetime
 
 from spaceone.core.manager import BaseManager
+from spaceone.core import utils
+from spaceone.core.error import *
 from spaceone.inventory.model.cloud_service_model import CloudService
 from spaceone.inventory.lib.resource_manager import ResourceManager
 from spaceone.inventory.manager.collection_state_manager import CollectionStateManager
+from spaceone.inventory.manager.identity_manager import IdentityManager
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -82,17 +85,32 @@ class CloudServiceManager(BaseManager, ResourceManager):
 
         return self.cloud_svc_model.get(**kwargs)
 
-    def list_cloud_services(self, query, target=None):
+    def list_cloud_services(self, query, target=None, change_filter=False, domain_id=None):
+        if change_filter:
+            query = self._change_project_group_filter(query, domain_id)
+            query = self._change_filter_tags(query)
+            query = self._change_only_tags(query)
+            query = self._change_sort_tags(query)
+
         # Append Query for DELETED filter (Temporary Logic)
         query = self._append_state_query(query)
         return self.cloud_svc_model.query(**query, target=target)
 
-    def analyze_cloud_services(self, query):
+    def analyze_cloud_services(self, query, change_filter=False, domain_id=None):
+        if change_filter:
+            query = self._change_project_group_filter(query, domain_id)
+            query = self._change_filter_tags(query)
+
         # Append Query for DELETED filter (Temporary Logic)
         query = self._append_state_query(query)
         return self.cloud_svc_model.analyze(**query)
 
-    def stat_cloud_services(self, query):
+    def stat_cloud_services(self, query, change_filter=False, domain_id=None):
+        if change_filter:
+            query = self._change_project_group_filter(query, domain_id)
+            query = self._change_filter_tags(query)
+            query = self._change_distinct_tags(query)
+
         # Append Query for DELETED filter (Temporary Logic)
         query = self._append_state_query(query)
         return self.cloud_svc_model.stat(**query)
@@ -179,3 +197,145 @@ class CloudServiceManager(BaseManager, ResourceManager):
             query['filter'].append(state_default_filter)
 
         return query
+
+    def _change_project_group_filter(self, query, domain_id):
+        change_filter = []
+
+        project_group_query = {
+            'filter': [],
+            'only': ['project_group_id']
+        }
+
+        for condition in query.get('filter', []):
+            key = condition.get('key', condition.get('k'))
+            value = condition.get('value', condition.get('v'))
+            operator = condition.get('operator', condition.get('o'))
+
+            if not all([key, operator]):
+                raise ERROR_DB_QUERY(reason='filter condition should have key, value and operator.')
+
+            if key == 'project_group_id':
+                project_group_query['filter'].append(condition)
+            else:
+                change_filter.append(condition)
+
+        if len(project_group_query['filter']) > 0:
+            identity_mgr: IdentityManager = self.locator.get_manager('IdentityManager')
+            response = identity_mgr.list_project_groups(project_group_query, domain_id)
+            project_group_ids = []
+            project_ids = []
+            for project_group_info in response.get('results', []):
+                project_group_ids.append(project_group_info['project_group_id'])
+
+            for project_group_id in project_group_ids:
+                response = identity_mgr.list_projects_in_project_group(project_group_id, domain_id, True,
+                                                                       {'only': ['project_id']})
+                for project_info in response.get('results', []):
+                    if project_info['project_id'] not in project_ids:
+                        project_ids.append(project_info['project_id'])
+
+            change_filter.append({'k': 'project_id', 'v': project_ids, 'o': 'in'})
+
+        query['filter'] = change_filter
+        return query
+
+    def _change_filter_tags(self, query):
+        change_filter = []
+
+        for condition in query.get('filter', []):
+            key = condition.get('k', condition.get('key'))
+            value = condition.get('v', condition.get('value'))
+            operator = condition.get('o', condition.get('operator'))
+
+            if key.startswith('tags.'):
+                hashed_key = self._get_hashed_key(key)
+
+                change_filter.append({
+                    'key': hashed_key,
+                    'value': value,
+                    'operator': operator
+                })
+
+            else:
+                change_filter.append(condition)
+        query['filter'] = change_filter
+        return query
+
+    def _change_only_tags(self, query):
+        change_only_tags = []
+        if 'only' in query:
+            for key in query.get('only', []):
+                if key.startswith('tags.'):
+                    hashed_key = self._get_hashed_key(key, only=True)
+                    change_only_tags.append(hashed_key)
+                else:
+                    change_only_tags.append(key)
+            query['only'] = change_only_tags
+
+        return query
+
+    def _change_distinct_tags(self, query):
+        if 'distinct' in query:
+            distinct_key = query['distinct']
+            if distinct_key.startswith('tags.'):
+                hashed_key = self._get_hashed_key(distinct_key)
+                query['distinct'] = hashed_key
+
+        return query
+
+    def _change_sort_tags(self, query):
+        if 'sort' in query:
+            if 'keys' in query['sort']:
+                change_filter = []
+                sort_keys = query['sort'].get('keys', [])
+                for condition in sort_keys:
+                    sort_key = condition.get('key', '')
+                    desc = condition.get('desc', False)
+
+                    if sort_key.startswith('tags.'):
+                        hashed_key = self._get_hashed_key(sort_key)
+
+                        change_filter.append({
+                            'key': hashed_key,
+                            'desc': desc
+                        })
+
+                    else:
+                        change_filter.append({
+                            'key': sort_key,
+                            'desc': desc
+                        })
+                query['sort']['keys'] = change_filter
+
+            elif 'key' in query['sort']:
+                change_filter = {}
+                sort_key = query['sort']['key']
+                desc = query['sort'].get('desc', False)
+
+                if sort_key.startswith('tags.'):
+                    hashed_key = self._get_hashed_key(sort_key)
+
+                    change_filter.update({
+                        'key': hashed_key,
+                        'desc': desc
+                    })
+
+                else:
+                    change_filter.update({
+                        'key': sort_key,
+                        'desc': desc
+                    })
+                query['sort'] = change_filter
+        return query
+
+    @staticmethod
+    def _get_hashed_key(key, only=False):
+        if key.count('.') < 2:
+            return key
+
+        prefix, provider, key = key.split('.', 2)
+        hash_key = utils.string_to_hash(key)
+        if only:
+            return f'{prefix}.{provider}.{hash_key}'
+        else:
+            return f'{prefix}.{provider}.{hash_key}.value'
